Added ES6 syntax
