+++ index.html
@@ -54,7 +54,6 @@
     <footer>
       <p>copyright...</p>
     </footer>
-    <script src="script.js"></script>
-    <script src="./navigation.js"></script>
+    <script type="module" src="script.js"></script>
   </body>
 </html>
